# Speed up tensor.resize_(sizes) when tensor has correct size