--- a/aten/src/ATen/native/Resize.cpp
+++ b/aten/src/ATen/native/Resize.cpp
@@ -9,30 +9,17 @@ Tensor& resize__cpu(Tensor& self, IntList size) {
   }
 
   auto* self_ = self.unsafeGetTensorImpl();
-
-  int totalSize = 1;
-  int nDimension = size.size();
-  if(nDimension != self_->dim()) {
-    self_->resize_dim(nDimension);
-  }
-
-  for (int d = nDimension - 1; d >= 0; d--) {
-    self_->set_size(d, size[d]);
-    if(d == nDimension - 1) {
-      self_->set_stride(d, 1);
-    } else {
-      // Keep stride monotonically increasing to match NumPy.
-      self_->set_stride(d, std::max<int64_t>(self_->size(d + 1), 1) * self_->stride(d + 1));
-    }
-    totalSize += (self_->size(d) - 1) * self_->stride(d);
-  }
+  self_->set_sizes_contiguous(size);
+  int totalSize = self_->numel();
 
   if (totalSize + self_->storage_offset() > 0) {
     if (!THTensor_getStoragePtr(self_)) {
       THTensor_stealAndSetStoragePtr(self_, THStorage_new(self_->dtype()));
     }
-    if (totalSize + self_->storage_offset() > THTensor_getStoragePtr(self_)->numel()) {
-      THStorage_resize(THTensor_getStoragePtr(self_), totalSize+self_->storage_offset());
+    if (totalSize + self_->storage_offset() > self_->storage().numel()) {
+      THStorage_resize(
+          THTensor_getStoragePtr(self_),
+          totalSize + self_->storage_offset());
     }
   }
   self_->maybe_zero_dim(size.size() == 0);