--- a/aten/src/ATen/native/cuda/Resize.cuh
+++ b/aten/src/ATen/native/cuda/Resize.cuh
@@ -1,0 +1,57 @@
+#pragma once
+
+#include "ATen/ATen.h"
+#include "THC/THCTensor.hpp"
+
+namespace at { namespace native {
+
+// These functions are called by native::resize_ as well as (legacy) THC resize.
+// They are not in THC/THCTensor.cpp because the at namespace is easier
+// to benchmark than THC; I can't get gbenchmark to call fns from THTensor.cpp
+
+static inline void maybe_resize_storage_cuda(TensorImpl* self, int64_t new_size) {
+  if (new_size + self->storage_offset() > 0) {
+    if (!THTensor_getStoragePtr(self)) {
+      AT_ERROR("Tensor: invalid null storage");
+    }
+    if (new_size + self->storage_offset() > self->storage().numel()) {
+      THCStorage_resize(
+          globalContext().getTHCState(),
+          THTensor_getStoragePtr(self),
+          new_size + self->storage_offset());
+    }
+  }
+}
+
+inline TensorImpl* resize_impl_cuda_(
+    TensorImpl* self,
+    IntList size,
+    c10::optional<IntList> stride,
+    bool device_guard=true) {
+  if (self->sizes() == size && (!stride || self->strides() == stride)) {
+    return self;
+  }
+
+  // NB: We don't need to hold the device guard when calling from TH
+  c10::optional<DeviceGuard> guard;
+  if (device_guard) {
+    guard = DeviceGuard(self->storage().device().index());
+  }
+
+  size_t storage_size = 1;
+  if (stride) {
+    self->set_sizes_and_strides(size, *stride);
+    // NB: storage size can be different from numel.
+    for (size_t dim = 0; dim < size.size(); ++dim) {
+      storage_size += (size[dim] - 1) * stride.value()[dim];
+    }
+  } else {
+    self->set_sizes_contiguous(size);
+    storage_size = self->numel();
+  }
+  maybe_resize_storage_cuda(self, self->numel());
+
+  return self;
+}
+
+}}