--- a/aten/src/THC/THCTensor.cpp
+++ b/aten/src/THC/THCTensor.cpp
@@ -9,6 +9,8 @@
 
 #include "THCTensorInfo.cuh"
 
+#include "ATen/native/cuda/Resize.cuh"
+
 int THCTensor_nDimension(THCState *state, const THCTensor *self) {
   return THTensor_nDimension(self);
 }
@@ -98,61 +100,12 @@ void THCTensor_resizeAs(THCState *state, THCTensor *self, THCTensor *src) {
 void THCTensor_resizeNd(THCState *state, THCTensor *self, int nDimension, const int64_t *size, const int64_t *stride)
 {
   AT_CHECK(nDimension >= 0, "resizeNd nDimension must be non-negative");
-  int d;
-  ptrdiff_t totalSize;
-  bool hascorrectsize = true;
-
-  for(d = 0; d < nDimension; d++)
-  {
-    if((self->dim() > d) && (size[d] != self->size(d))) {
-      hascorrectsize = false;
-    }
-
-    // NB: this used to test that stride[d] was >= 0
-    if((self->dim() > d) && stride && (stride[d] != self->stride(d))) {
-      hascorrectsize = false;
-    }
+  at::IntList sizes(size, nDimension);
+  at::optional<at::IntList> strides;
+  if (stride) {
+    strides = at::IntList(stride, nDimension);
   }
-
-  if(nDimension != self->dim()) {
-    hascorrectsize = false;
-  }
-
-  if(hascorrectsize) {
-    return;
-  }
-
-  if(nDimension != self->dim())
-  {
-    self->resize_dim(nDimension);
-  }
-
-  totalSize = 1;
-  for(d = nDimension-1; d >= 0; d--)
-  {
-    self->set_size(d, size[d]);
-    if(stride && (stride[d] >= 0) ) {
-      self->set_stride(d, stride[d]);
-    } else {
-      if(d == nDimension-1) {
-        self->set_stride(d, 1);
-      } else {
-        // Keep stride monotonically increasing to match NumPy.
-        self->set_stride(d, std::max<int64_t>(self->size(d+1),1)*self->stride(d+1));
-      }
-    }
-    totalSize += (self->size(d)-1)*self->stride(d);
-  }
-
-  if(totalSize+self->storage_offset() > 0)
-  {
-    if(!THTensor_getStoragePtr(self)) {
-      THError("Tensor: invalid null storage");
-    }
-    if(totalSize+self->storage_offset() > THTensor_getStoragePtr(self)->numel()) {
-      THCStorage_resize(state, THTensor_getStoragePtr(self), totalSize+self->storage_offset());
-    }
-  }
+  at::native::resize_impl_cuda_(self, sizes, strides, /*device_guard=*/false);
 }
 
 void THCTensor_set(THCState *state, THCTensor *self, THCTensor *src)