@zou3519 2018-10-23
@ezyang I figured out a compromise for (3): I've replaced the old TH implementation with the new implementation in this PR. I modified `TensorImpl::set_sizes_and_strides` to match the old `THTensor_(resizeNd)` semantics -- could you take a look and see if this is okay?