@bddppq 2018-10-26
This PR has broken PyTorch ROCm tests:
https://ci.pytorch.org/jenkins/job/pytorch-builds/job/py2-clang7-rocmdeb-ubuntu16.04-test/1261//console

cc @iotamudelta 

```
06:33:27 test_zeros_like (__main__.TestUncoalescedSparse) ... ok
06:33:28 
06:33:28 ======================================================================
06:33:28 ERROR: test_add_dense_sparse_mismatch (__main__.TestCudaSparse)
06:33:28 ----------------------------------------------------------------------
06:33:28 Traceback (most recent call last):
06:33:28   File "test_sparse.py", line 933, in test_add_dense_sparse_mismatch
06:33:28     test_shape([3, 4, 0], [1, 4], [4, 4, 4, 0], [3, 4, 4, 0])
06:33:28   File "test_sparse.py", line 923, in test_shape
06:33:28     x = torch.zeros(dense_size, dtype=self.value_dtype, device=self.device)
06:33:28 RuntimeError: cuda runtime error (1011) : hipErrorInvalidValue at /var/lib/jenkins/workspace/aten/src/THC/generic/THCTensorMath.cu:25
06:33:28 
06:33:28 ======================================================================
06:33:28 ERROR: test_add_dense_sparse_mismatch (__main__.TestCudaUncoalescedSparse)
06:33:28 ----------------------------------------------------------------------
06:33:28 Traceback (most recent call last):
06:33:28   File "test_sparse.py", line 933, in test_add_dense_sparse_mismatch
06:33:28     test_shape([3, 4, 0], [1, 4], [4, 4, 4, 0], [3, 4, 4, 0])
06:33:28   File "test_sparse.py", line 923, in test_shape
06:33:28     x = torch.zeros(dense_size, dtype=self.value_dtype, device=self.device)
06:33:28 RuntimeError: cuda runtime error (1011) : hipErrorInvalidValue at /var/lib/jenkins/workspace/aten/src/THC/generic/THCTensorMath.cu:25
06:33:28 
06:33:28 ======================================================================
06:33:28 ERROR: test_reduction_empty (test_torch.TestTorch)
06:33:28 ----------------------------------------------------------------------
06:33:28 Traceback (most recent call last):
06:33:28   File "/var/lib/jenkins/workspace/test/test_torch.py", line 1047, in test_reduction_empty
06:33:28     self.assertEqual(torch.empty((2, 0, 1), device=device), fn(x, dim=2, keepdim=True))
06:33:28   File "/var/lib/jenkins/workspace/test/test_torch.py", line 1024, in <lambda>
06:33:28     ('sum', lambda *args, **kwargs: torch.sum(*args, **kwargs), 0),
06:33:28 RuntimeError: cuda runtime error (1011) : hipErrorInvalidValue at /var/lib/jenkins/workspace/aten/src/THC/generic/THCTensorMath.cu:25
06:33:28 
06:33:28 ----------------------------------------------------------------------
```