@jithunnair-amd 2018-10-26
This looks like more cases of "hipMemset with size 0"