@iotamudelta 2018-10-26
so please skip for now, this will be fixed with `ROCm 1.9.2`